requirements.txt
# Core
torch>=2.0.0
transformers>=4.30.0
ftfy
tqdm
numpy
pillow
faiss-cpu
scikit-learn
sqlite-utils
streamlit
python-multipart
opencv-python
pandas
